requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2.5
django-cache-url==3.2.3
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-js-asset==1.2.2
gunicorn==20.1.0
marshmallow==3.13.0
Pillow==8.3.1
psycopg2==2.9.1
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
